Made 2 different components
One fo rmath operations and one for digits
